Refactor GUI Classes
