+++ GUI/src/main/java/GUI/PlaylistPanel.java
@@ -1,0 +1,16 @@
+package GUI;
+
+
+import javax.swing.*;
+import javax.imageio.ImageIO;
+import service.Extract;
+import java.util.ArrayList;
+import java.awt.*;
+import java.awt.event.*;
+import java.io.File;
+import java.util.Vector;
+
+
+public class PlaylistPanel extends JPanel {
+
+}
